+++ src/main/java/org/n52/sta/edm/provider/entities/LocationEntityProvider.java
@@ -72,7 +72,7 @@ public class LocationEntityProvider extends AbstractSensorThingsEntityProvider {
     protected CsdlEntityType createEntityType() {
         //create EntityType primitive properties
         CsdlProperty id = new CsdlProperty().setName(ID_ANNOTATION)
-                .setType(EdmPrimitiveTypeKind.Int32.getFullQualifiedName())
+                .setType(EdmPrimitiveTypeKind.Int64.getFullQualifiedName())
                 .setNullable(false);
         CsdlProperty name = new CsdlProperty().setName(PROP_NAME)
                 .setType(EdmPrimitiveTypeKind.String.getFullQualifiedName())
